requirements.txt
Django~=3.1.7
django-filter~=2.4.0
djangorestframework~=3.12.2
djangorestframework-simplejwt~=4.6.0
pillow~=8.1.2
pytest~=6.2.2
model-bakery~=1.2.1